fix: Resolve CodeQL alerts in src tier
Fixes 46 of 183 CodeQL alerts across src/ runtime libraries; the
remaining 137 are deliberately dismissed with per-alert rationale in
dismissals.tsv (false positive / won't fix / used in tests).

- catch-of-all-exceptions: type SaveManager and EncryptedPersistenceApi
  save-slot probe catches (IOException, UnauthorizedAccessException,
  InvalidDataException, EndOfStreamException, JsonException); narrow
  TestBridgeClient message parsing to JsonException, EntitySnapshot
  conversion to InvalidCastException/FormatException/OverflowException,
  UdpTransport disconnect send to SocketException/ObjectDisposedException,
  and UIColorPicker hex parsing to FormatException/OverflowException/
  ArgumentException. Bridge/IPC dispatch, logging, hot-reload and
  background-worker containment boundaries are kept and documented.
- loss-of-precision: float-typed arithmetic in WavDecoder,
  SilkAudioContext, AudioClipAsset, GridCoordinate, SilkTextRenderer,
  UdpTransport, ParallelProfiler; FleeAction's intentional integer
  centering made explicit via an int local.
- local-shadows-member: rename shadowing locals in GraphRenderSystem;
  rename Silk2DRenderer's projection field to screenProjection and
  EnsureCapacity parameters (interface parameter names kept for S927).
- constant-condition: drop redundant checksum conjunct in Ghost/Replay
  file formats, hoist device null check out of PauseAll loop, and
  replace the always-false ownsEntity/existingTypes conjunct in
  NetworkServerPlugin owner-state validation.
- useless-upcast/cast-to-self: assign World directly to capability
  fields in StateControllerImpl; drop redundant collection-expression
  casts in AssetManifest/AssetEntry.
- complex-condition: pattern-match Unicode range checks in text shapers.
- inefficient-containskey: TryGetValue in LogAssertions.
- equality-on-floats, obsolete prefab mock calls, ref-parameter counts:
  dismissed as exact-by-construction, test-only mock of a deprecated
  API, and interface-mandated/hot-path signatures respectively.

+++ src/KeenEyes.Navigation.Grid/GridCoordinate.cs
@@ -123,8 +123,8 @@ public readonly record struct GridCoordinate(int X, int Y)
     /// <returns>The straight-line distance.</returns>
     public float EuclideanDistance(GridCoordinate other)
     {
-        int dx = X - other.X;
-        int dy = Y - other.Y;
+        float dx = X - other.X;
+        float dy = Y - other.Y;
         return MathF.Sqrt(dx * dx + dy * dy);
     }
